ran clippy fix

+++ src/algorithms/miller_rabin.rs
@@ -175,7 +175,7 @@ where
     let mut rng = oorandom::Rand64::new(ZZ.default_hash(n) as u128);
     let mut n_minus_one = ZZ.sub_ref_fst(n, ZZ.one());
     let s = ZZ.abs_lowest_set_bit(&n_minus_one).unwrap();
-    ZZ.euclidean_div_pow_2(&mut n_minus_one, s as usize);
+    ZZ.euclidean_div_pow_2(&mut n_minus_one, s);
     let d = n_minus_one;
 
     for _i in 0..k {
